insights/requirements.txt
pandas
python-dateutil
flask-cors
